pyproject.toml
[tool.poetry]
name = "canarytokens"
version = "0.3.0"
description = "Canarytokens"
authors = ["Thinkst Applied Research"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.9"
pydantic = {extras = ["dotenv", "email"], version = "^1.9.0"}
PyQRCode = "^1.2.1"
redis = "^4.2.0"
httpx = "^0.23.0"
Jinja2 = "^3.1.2"
requests = "^2.27.1"
python-multipart = "^0.0.9"
pyOpenSSL = "^23.0.0"
service-identity = "^21.1.0"
PyYAML = "^6.0"
PyNaCl = "^1.5.0"
segno = "^1.5.2"
sendgrid = "^6.9.7"
minify-html = "^0.9.2"
kafka-python = "^2.2.13"

# A list of all of the optional dependencies, some of which are included in the
# below `extras`. They can be opted into by apps.
fastapi = { version = "^0", optional = true }
sentry-sdk = {extras = ["fastapi"], version = "^1.9.5", optional = true}
uvicorn = { version = "^0.17.6", optional = true }
Twisted = {version = "^22.4.0", optional = true}
faker = "^17.6.0"
azure-core = "^1.23.1"
azure-identity = "^1.10.0"
cssutils = "^1.0.2"
advocate = {git = "https://github.com/JordanMilne/Advocate.git"}
pycountry-convert = "^0.7.2"

[tool.poetry.extras]
web = ["fastapi", "uvicorn", "sentry-sdk"]
twisted = ["Twisted", "sentry-sdk"]

[tool.poetry.dev-dependencies]
pytest = "^7.1.0"
black = "^22.1.0"
pre-commit = "^2.17.0"
mypy = "^0.941"
coverage = "^7.0.0"
requests = "^2.27.1"
hypothesis = "^6.39.3"
pyngrok = "^6.0.0"
rpdb = "^0.1.6"
toml = "^0.10.2"
types-toml = "^0.10.7"
types-pyOpenSSL = "^22.0.4"
types-PyYAML = "^6.0.9"
types-redis = "^4.1.18"
mypy-zope = "^0.3.6"
pytest-twisted = "^1.13.4"
pytest-asyncio = "^0.18.3"
types-requests = "^2.27.15"
Pillow = "^10.0.1"
pyzbar = "^0.1.9"
debugpy = "^1.6.0"
pytest-memray = {version = "^1.0.0", platform = "linux"}
python-docx = "^0.8.11"
boto3 = "^1.23.10"
boto3-stubs = {extras = ["essential"], version = "^1.24.17"}
botocore = "^1.34.141"
modernize = "^0.8.0"
deepdiff = "^5.8.1"
pypdf = "^4.2.0"
easywebdav = "1.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
